CLEAN refactor
Moved to a more CLEAN approach
