+++ app/src/main/AndroidManifest.xml
@@ -19,7 +19,7 @@
         tools:targetApi="32">
 
         <activity
-            android:name="com.spacex.usecases.launch.LaunchActivity"
+            android:name="com.spacex.ui.launch.LaunchActivity"
             android:exported="true"
             android:label="@string/app_name"
             android:theme="@style/Theme.SpaceX">
@@ -29,7 +29,7 @@
             </intent-filter>
         </activity>
 
-        <activity android:name="com.spacex.usecases.main.MainActivity"></activity>
+        <activity android:name="com.spacex.ui.main.MainActivity"></activity>
     </application>
 
 </manifest>
