+++ app/src/main/java/com/spacex/UseCaseResponse.kt
@@ -1,0 +1,8 @@
+package com.spacex
+
+sealed class UseCaseResponse<out T : Any> {
+    class Success<T : Any>(val data: T) : UseCaseResponse<T>()
+    object NoData : UseCaseResponse<Nothing>()
+    object NoNetwork : UseCaseResponse<Nothing>()
+    object GenericError : UseCaseResponse<Nothing>()
+}
