+++ app/src/main/java/com/spacex/ViewModelResponse.kt
@@ -1,0 +1,7 @@
+package com.spacex
+
+enum class ViewModelResponse {
+    NULL_EMPTY_DATA,
+    NO_NETWORK,
+    GENERIC_ERROR
+}
