+++ app/src/main/java/com/spacex/data/model/Launch.kt
@@ -1,0 +1,29 @@
+package com.spacex.data.model
+
+import java.io.Serializable
+import java.util.*
+
+data class Launch(
+
+    var name: String,
+    var date_utc: Date,
+    var success: Boolean,
+    var links: Links
+
+) : Serializable
+
+data class Links(
+
+    var presskit: String?,
+    var youtube_id: String?,
+    var wikipedia: String?
+
+) : Serializable
+
+data class LaunchesQueryResponse(
+
+    var docs: MutableList<Launch>
+
+) : Serializable
+
+
