+++ app/src/main/java/com/spacex/data/model/Query.kt
@@ -1,0 +1,15 @@
+package com.spacex.data.model
+
+import java.io.Serializable
+
+data class Query(
+
+    var query: QueryObject
+
+) : Serializable
+
+data class QueryObject(
+
+    var rocket: String
+
+) : Serializable
