+++ app/src/main/java/com/spacex/data/repository/RepositoryResponse.kt
@@ -1,0 +1,10 @@
+package com.spacex.data.repository
+
+sealed class RepositoryResponse<out T : Any>{
+
+    class Success<T: Any>(val data: T) : RepositoryResponse<T>()
+    object NoData : RepositoryResponse<Nothing>()
+    object NoNetwork : RepositoryResponse<Nothing>()
+    object GenericError : RepositoryResponse<Nothing>()
+
+}
