--- app/src/main/java/com/spacex/model/data/Query.kt
@@ -1,15 +1,0 @@
-package com.spacex.model.data
-
-import java.io.Serializable
-
-data class Query(
-
-    var query: QueryObject
-
-) : Serializable
-
-data class QueryObject(
-
-    var rocket: String
-
-) : Serializable
