--- app/src/main/java/com/spacex/util/NetworkError.kt
@@ -1,7 +1,0 @@
-package com.spacex.util
-
-enum class NetworkError {
-    NULL_EMPTY_DATA,
-    NO_NETWORK,
-    GENERIC_ERROR
-}
