+++ app/src/main/res/layout/activity_launch.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".usecases.launch.LaunchActivity">
+    tools:context=".ui.launch.LaunchActivity">
 
     <LinearLayout
         android:layout_width="wrap_content"
