+++ app/src/main/res/layout/fragment_rocket.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".usecases.main.RocketFragment">
+    tools:context=".ui.main.RocketFragment">
 
     <LinearLayout
         android:id="@+id/linearLoading"
