+++ app/src/main/res/layout/fragment_rocket_list.xml
@@ -5,7 +5,7 @@
     android:id="@+id/swipeRefresh"
     android:layout_width="match_parent"
     android:layout_height="wrap_content"
-    tools:context=".usecases.main.RocketListFragment">
+    tools:context=".ui.main.RocketListFragment">
 
     <RelativeLayout
         android:layout_width="match_parent"
